requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
pymongo==4.8.0
Flask-PyMongo==2.3.0
dnspython==2.6.1
python-dotenv==1.0.1
bcrypt==4.1.2
Flask-WTF==1.2.1
WTForms==3.1.2
gunicorn==21.2.0
